+++ src/ZToolLib/ztl_unit_test.c
@@ -19,7 +19,7 @@ char* ZuStrAlloc(int size)
 
 char* ZuStrCopy(const char* old)
 {
-    int len = strlen(old);
+    int len = (int)strlen(old);
     char* newStr = ZuStrAlloc(len + 1);
     strcpy(newStr, old);
     return newStr;
@@ -61,7 +61,7 @@ void ZuStringAppend(ZuString* str, const char* text)
         text = "NULL";
     }
 
-    length = strlen(text);
+    length = (int)strlen(text);
     if (str->length + length + 1 >= str->size)
         ZuStringResize(str, str->length + length + 1 + STRING_INC);
     str->length += length;
@@ -88,7 +88,7 @@ void ZuStringAppendFormat(ZuString* str, const char* format, ...)
 
 void ZuStringInsert(ZuString* str, const char* text, int pos)
 {
-    int length = strlen(text);
+    int length = (int)strlen(text);
     if (pos > str->length)
         pos = str->length;
     if (str->length + length + 1 >= str->size)
